requirements.txt
accelerate==1.0.0
adafruit-nrfutil==0.5.3.post16
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
distro==1.9.0
ecdsa==0.19.0
et-xmlfile==1.1.0
factor_analyzer==0.5.1
filelock==3.16.1
fonttools==4.53.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.25.2
idna==3.7
isort==5.13.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
krippendorff==0.7.0
MarkupSafe==3.0.1
matplotlib==3.9.1
mccabe==0.7.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4
numpy==2.0.0
ollama==0.3.3
openai==1.36.0
openpyxl==3.1.5
pandas==2.2.2
pathspec==0.12.1
patsy==0.5.6
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pylint==3.2.6
pyparsing==3.1.2
pyserial==3.5
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.0
sniffio==1.3.1
statsmodels==0.14.2
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.1
tomlkit==0.13.0
torch==2.4.1
tqdm==4.66.4
transformers==4.45.2
tzdata==2024.1
urllib3==2.2.3
